LayoutTests/platform/blackberry/fast/dom/DeviceMotion/no-page-cache-expected.txt

CONSOLE MESSAGE: line 4: TypeError: Result of expression 'window.opener.reportPageOneOnload' [undefined] is not a function.
FAIL: Timed out waiting for notifyDone to be called
